+++ internal/displacement/displacement.go
@@ -24,12 +24,12 @@ type minimalTriangle struct {
 	// Material
 	material material.Material
 	// Texture coordinates
-	u0 float64
-	u1 float64
-	u2 float64
-	v0 float64
-	v1 float64
-	v2 float64
+	u0 float32
+	u1 float32
+	u2 float32
+	v0 float32
+	v1 float32
+	v2 float32
 }
 
 // tessellate splits a triangle in four smaller triangles.
@@ -101,7 +101,7 @@ func tessellate(in *minimalTriangle) []*minimalTriangle {
 }
 
 // getDisplacementVariation calculates the variation in displacement values at the triangle vertices
-func getDisplacementVariation(tri *minimalTriangle, displacementMap texture.Texture) float64 {
+func getDisplacementVariation(tri *minimalTriangle, displacementMap texture.Texture) float32 {
 	// Sample displacement at each vertex (we only care about the Z component)
 	d0 := displacementMap.Value(tri.u0, tri.v0, nil).Z
 	d1 := displacementMap.Value(tri.u1, tri.v1, nil).Z
@@ -114,7 +114,7 @@ func getDisplacementVariation(tri *minimalTriangle, displacementMap texture.Text
 	return maxDisp - minDisp
 }
 
-func isTessellatedEnough(tri *minimalTriangle, maxDeltaU float64, maxDeltaV float64, displacementMap texture.Texture, min, max, adaptiveThreshold float64) bool {
+func isTessellatedEnough(tri *minimalTriangle, maxDeltaU float32, maxDeltaV float32, displacementMap texture.Texture, min, max, adaptiveThreshold float32) bool {
 	// Check UV coordinate constraints (minimum tessellation)
 	uvTessellated := math.Abs(tri.u1-tri.u0) <= maxDeltaU &&
 		math.Abs(tri.u2-tri.u1) <= maxDeltaU &&
@@ -142,7 +142,7 @@ func isTessellatedEnough(tri *minimalTriangle, maxDeltaU float64, maxDeltaV floa
 
 // ApplyDisplacementMap tessellates the triangles and applies the displacement map to all of them.
 // It uses adaptive tessellation that only subdivides areas with significant displacement variation.
-func ApplyDisplacementMap(triangles []*hitable.Triangle, displacementMap texture.Texture, min, max float64) ([]*hitable.Triangle, error) {
+func ApplyDisplacementMap(triangles []*hitable.Triangle, displacementMap texture.Texture, min, max float32) ([]*hitable.Triangle, error) {
 	var resU, resV int
 
 	if displacementTexture, ok := displacementMap.(*texture.ImageTxt); !ok {
@@ -172,8 +172,8 @@ func ApplyDisplacementMap(triangles []*hitable.Triangle, displacementMap texture
 	// Use coarser UV tessellation limits (4x the texture pixel size)
 	// This allows adaptive tessellation to take over for detail
 	tessellationFactor := 4.0
-	maxDeltaU := tessellationFactor / float64(resU-1)
-	maxDeltaV := tessellationFactor / float64(resV-1)
+	maxDeltaU := tessellationFactor / float32(resU-1)
+	maxDeltaV := tessellationFactor / float32(resV-1)
 
 	// Adaptive threshold: stop tessellating when displacement variation is less than this
 	// This is in world space units. Smaller values = more tessellation for finer detail.
@@ -185,7 +185,7 @@ func ApplyDisplacementMap(triangles []*hitable.Triangle, displacementMap texture
 	return applyDisplacement(tessellated, displacementMap, min, max), nil
 }
 
-func applyTessellation(in []*minimalTriangle, maxDeltaU float64, maxDeltaV float64, displacementMap texture.Texture, min, max, adaptiveThreshold float64) []*minimalTriangle {
+func applyTessellation(in []*minimalTriangle, maxDeltaU float32, maxDeltaV float32, displacementMap texture.Texture, min, max, adaptiveThreshold float32) []*minimalTriangle {
 	out := []*minimalTriangle{}
 
 	log.Info("Applying adaptive tessellation")
@@ -216,7 +216,7 @@ func applyTessellation(in []*minimalTriangle, maxDeltaU float64, maxDeltaV float
 	}
 }
 
-func applyDisplacement(in []*minimalTriangle, displacementMap texture.Texture, min, max float64) []*hitable.Triangle {
+func applyDisplacement(in []*minimalTriangle, displacementMap texture.Texture, min, max float32) []*hitable.Triangle {
 	displaced := []*hitable.Triangle{}
 
 	log.Info("Applying displacement")
